# contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_case_document_api_v1_supply_chain_po_cases__case_id__documents_post": {
        "properties": {
          "doc_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "doc_type",
          "file"
        ],
        "title": "Body_upload_case_document_api_v1_supply_chain_po_cases__case_id__documents_post",
        "type": "object"
      },
...
      "CaseDocumentView": {
        "description": "A document as the API shows it. The object key stays on the server.",
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "doc_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "po_case_id": {
            "format": "uuid",
            "title": "Po Case Id",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "uploaded_at": {
            "format": "date-time",
            "title": "Uploaded At",
            "type": "string"
          },
          "uploaded_by": {
            "format": "uuid",
            "title": "Uploaded By",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "po_case_id",
          "doc_type",
          "filename",
          "content_type",
          "size_bytes",
          "sha256",
          "version",
          "uploaded_by",
          "uploaded_at"
        ],
        "title": "CaseDocumentView",
        "type": "object"
      },
...
      "DocumentType": {
        "description": "ADR 0021's fourteen document types, from process.md section 2.",
        "enum": [
          "proposal_list",
          "product_image",
          "sample_photo",
          "sample_evaluation",
          "sample_revision_request",
          "product_profile_bm04",
          "official_item_code",
          "supplier_confirmation_email",
          "purchase_order",
          "deposit_docs",
          "payment_docs",
          "packaging_content",
          "user_manual",
          "maquette"
        ],
        "title": "DocumentType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Digital Worker Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/supply-chain/documents/{document_id}/content": {
      "get": {
        "operationId": "download_case_document_api_v1_supply_chain_documents__document_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The file, as an attachment of its stored type."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Case Document",
        "tags": [
          "supply_chain"
        ]
      }
    },
...
    "/api/v1/supply-chain/po-cases/{case_id}/documents": {
      "get": {
        "operationId": "list_case_documents_api_v1_supply_chain_po_cases__case_id__documents_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CaseDocumentView"
                  },
                  "title": "Response List Case Documents Api V1 Supply Chain Po Cases  Case Id  Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Case Documents",
        "tags": [
          "supply_chain"
        ]
      },
      "post": {
        "operationId": "upload_case_document_api_v1_supply_chain_po_cases__case_id__documents_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "description": "Optional. Retrying with the same key returns the first response instead of acting twice; reusing it for a different request is a 409.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional. Retrying with the same key returns the first response instead of acting twice; reusing it for a different request is a 409.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_case_document_api_v1_supply_chain_po_cases__case_id__documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDocumentView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Case Document",
        "tags": [
          "supply_chain"
        ]
      }
    },
...
  }
}